requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beartype==0.11.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
et-xmlfile==1.1.0
fastjsonschema==2.16.2
Flask==2.2.2
Flask-Cors==3.0.10
flet==0.3.2
fonttools==4.38.0
idna==3.4
imageio==2.24.0
imbalanced-learn==0.10.1
imblearn==0.0
importlib-metadata==5.2.0
importlib-resources==5.10.1
install==1.3.5
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.3
jupyter-server==1.23.4
jupyter_client==7.4.8
jupyter_core==4.12.0
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.17.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.17.0
oauthlib==3.2.2
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==22.0
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
PyWavelets==1.3.0
pywin32==305
pywinpty==2.0.9
PyYAML==3.13
pyzmq==24.0.1
repath==0.9.0
requests==2.28.1
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
shapely==2.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tabulate==0.9.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tomli==2.0.1
torch==1.7.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typing_extensions==4.4.0
urllib3==1.26.13
watchdog==2.2.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wincertstore==0.2
xlrd==1.2.0
zipp==3.11.0
